Save augmented data in .npz format instead of .npy format
- It turned out that .npy files containing multiple objects are not
straightforward to read using C++/pybind11. The only way that I know
how to read such a file is using np.attr("load")(filename). In the
case of an .npy file this only reads the first object, whereas in the
case of an .npz file the entire content is read.