Added attempts to increase the efficiency of dense matrix products
During backpropagation dense matrix products of the form 'column
major' times 'row major' are computed. This can take twice as
long as the other three variants.

In this case it can be faster to change the layout of one of the
arguments. This is now available using the flags

NERVA_DENSE_PRODUCT1 use MKL's omat_copy to change the layout
NERVA_DENSE_PRODUCT2 use Eigen to change the layout

When NERVA_DENSE_PRODUCT1 is set, the training is between 3 and 4%
faster on a CIFAR10 experiment with 95% sparsity.

Note that in these experiments a copy is made of the first argument.
An experiment to change the layout in place using MKL's imat_copy
has failed. It turned out that imat_copy is an extremely inefficient
operation.