Use a direct MKL call instead of an Eigen matrix product
- Timing intermediate results revealed that the dense matrix product
`DW = DZ.transpose() * X;` in Eigen can be very expensive for the
case `row major` times `column major`. This caused the batch
normalization examples to run three times slower for the rowwise
versions of the tools.

This can be solved using a direct call to the MKL library. Since this
may happen in many more places, the dense matrix product needs to be
wrapped in a function.